Reorganize files to get basic web app loading
- Separate server code into server directory
- Put client code into wwwroot directory
- Get basic static assets node/express server working
- Fix numerous syntax errors in javascript
- Fix some of angular app initialization
- App still does not load entirely and socket.io code should probably start over
