+++ .gitignore
@@ -1,0 +1,1 @@
+server/node_modules
