--- Assets/Vendors/Javascript/app.js
@@ -1,14 +1,0 @@
-angular.module('NoteWrangler', ['ngRoute', 'ngResource']);
-
-var express = require('express');
-var app = express();
-var angular = angular();
-
-app.use(express.static('aws-portfolio'));
-}); 
-
-app.listen(5000, function(){
-  console.log('Listening on port 3000');
-});
-
-angular.module('NoteWrangler', ['ngRoute', 'ngResource']);
