--- Views/index.html
@@ -1,56 +1,0 @@
-<!DOCTYPE html>
-<html lang='en' ng-app='NoteWrangler'>
-<head>
-  <meta charset='UTF-8'>
-  <title>Music Icons</title>
-  <link rel='stylesheet' href='assets/styles/bootstrap.css'></link>
-<link rel='stylesheet' href='assets/styles/bootstrap-theme.css'></link>
-</head>
-<body>
-  <nav class='navbar navbar-default'>
-    <div class='container-fluid'>
-     <div class='navbar-header'>
-      <div class='navbar-brand'>
-        <a href='#/'>Concerts</a>
-      </div>
-    </div>
-    <mi-page-nav></mi-page-nav>
-  </div>
-  </nav>
-  <div ng-view class='container-fluid'></div>
-</body>
-
-<!-- Vendor -->
-<script src='assets/javascript/vendor/jquery-2.1.3.min.js'></script>
-<script src='assets/javascript/vendor/angular.js'></script>
-<script src='assets/javascript/vendor/angular-route.js'></script>
-<script src='assets/javascript/vendor/angular-resource.js'></script>
-
-<!-- App -->
-<script src='assets/javascript/app.js'></script>
-<script src='assets/javascript/routes.js'></script>
-
-<!-- Directives -->
-<script src='assets/javascript/controllers/directives/miPageNav.js'></script>
-<script src='assets/javascript/controllers/directives/miCategorySelect.js'></script>
-
-<!-- Controllers -->
-<script src='assets/javascript/controllers/notesIndexController.js'></script>
-<script src='assets/javascript/controllers/notesCreateController.js'></script>
-<script src='assets/javascript/controllers/notesDeleteController.js'></script>
-<script src='assets/javascript/controllers/notesEditController.js'></script>
-<script src='assets/javascript/controllers/notesShowController.js'></script>
-<script src='assets/javascript/controllers/userIndexController.js'></script>
-<script src='assets/javascript/controllers/userShowController.js'></script>
-
-<!-- Filters -->
-<script src='assets/javascript/filters/categoryFilter.js'></script>
-
-
-<!-- Services -->
-<script src='assets/javascript/Services/note.js'></script>
-<script src='assets/javascript/Services/server.js'></script>
-<script src='assets/javascript/Services/user.js'></script>
-<script src='assets/javascript/Services/categories.js'></script>
-
-</html>
