--- index.html
@@ -1,37 +1,0 @@
-<!DOCTYPE html>
-<html lang='en' ng-app='NoteWrangler'>
-<head>
-  <meta charset='UTF-8'>
-  <title>Music Icons</title>
-  <link rel='stylesheet' href='assets/styles/bootstrap.css'></link>
-<link rel='stylesheet' href='assets/styles/bootstrap-theme.css'></link>
-</head>
-<body>
-  <h1>Concerts</h1>
-  <div ng-view class='container-fluid'></div>
-</body>
-
-<!-- Vendor -->
-<script src='assets/javascript/vendor/jquery-2.1.3.min.js'></script>
-<script src='assets/javascript/vendor/angular.js'></script>
-<script src='assets/javascript/vendor/angular-route.js'></script>
-<script src='assets/javascript/vendor/angular-resource.js'></script>
-
-<!-- App -->
-<script src='assets/javascript/app.js'></script>
-<script src='assets/javascript/routes.js'></script>
-
-<!-- Directives -->
-
-<!-- Controllers -->
-<script src='assets/javascript/controllers/notesIndexController.js'></script>
-<script src='assets/javascript/controllers/notesCreateController.js'></script>
-<script src='assets/javascript/controllers/notesDeleteController.js'></script>
-<script src='assets/javascript/controllers/notesEditController.js'></script>
-<script src='assets/javascript/controllers/notesShowController.js'></script>
-
-<!-- Services -->
-<script src='assets/javascript/services/note.js'></script>
-<script src='assets/javascript/services/server.js'></script>
-
-</html>
