+++ server/package.json
@@ -1,0 +1,10 @@
+{
+  "name": "music-icons-concert",
+  "version": "1.0.0",
+  "main": "server.js",
+  "private": true,
+  "dependencies": {
+    "express": "4.15.2",
+    "socket.io": "1.7.4"
+  }
+}
