+++ server/server.js
@@ -1,0 +1,9 @@
+var path = require('path');
+var express = require('express');
+var app = express();
+
+app.use(express.static(path.join(__dirname, '../wwwroot')));
+
+app.listen(3000, function(){
+  console.log('Listening on port 3000');
+});
