+++ wwwroot/controllers/notesDeleteController.js
