+++ wwwroot/controllers/userIndexController.js
@@ -1,4 +1,4 @@
 angular.module('NoteWrangler').controller('UserIndexController', function(User, $scope){
   $scope.user = User.query();
-  $scope.search = ();
+  // $scope.search = ();
 });
