+++ wwwroot/index.html
@@ -1,0 +1,55 @@
+<!DOCTYPE html>
+<html lang='en' ng-app='NoteWrangler'>
+<head>
+  <meta charset='UTF-8'>
+  <title>Music Icons</title>
+  <link rel='stylesheet' href='/vendor/bootstrap.css'></link>
+<link rel='stylesheet' href='/vendor/bootstrap-theme.css'></link>
+</head>
+<body>
+  <nav class='navbar navbar-default'>
+    <div class='container-fluid'>
+     <div class='navbar-header'>
+      <div class='navbar-brand'>
+        <a href='#/'>Concerts</a>
+      </div>
+    </div>
+    <mi-page-nav></mi-page-nav>
+  </div>
+  </nav>
+  <div ng-view class='container-fluid'></div>
+</body>
+
+<!-- Vendor -->
+<script src='/vendor/jquery-2.1.3.min.js'></script>
+<script src='/vendor/angular.js'></script>
+<script src='/vendor/angular-route.js'></script>
+<script src='/vendor/angular-resource.js'></script>
+
+<!-- App -->
+<script src='/routes.js'></script>
+
+<!-- Directives -->
+<script src='/directives/miPageNav.js'></script>
+<script src='/directives/miCategorySelect.js'></script>
+
+<!-- Controllers -->
+<script src='/controllers/notesIndexController.js'></script>
+<script src='/controllers/notesCreateController.js'></script>
+<script src='/controllers/notesDeleteController.js'></script>
+<script src='/controllers/notesEditController.js'></script>
+<script src='/controllers/notesShowController.js'></script>
+<script src='/controllers/userIndexController.js'></script>
+<script src='/controllers/userShowController.js'></script>
+
+<!-- Filters -->
+<script src='/filters/categoryFilter.js'></script>
+
+
+<!-- Services -->
+<script src='/Services/note.js'></script>
+<script src='/Services/server.js'></script>
+<script src='/Services/user.js'></script>
+<script src='/Services/categories.js'></script>
+
+</html>
