+++ wwwroot/services/server.js
@@ -1,0 +1,5 @@
+var socket = io.connect();
+socket.on('answers', function (data) {
+  console.log(data);
+  socket.emit('gum', { my: 'data' });
+});
