+++ wwwroot/templates/Notes/show.html
